Fix errors resulted from rebase
