+++ client/src/browser_action/browser_action.html
@@ -1,23 +1,19 @@
 <!DOCTYPE html>
 <html>
-
-<head>
-  <style type="text/css">
-    body {
-      padding: 10px;
-      width: 500px;
-      height: 500px;
-    }
-  </style>
-
-  <script async src="popup.js">
-  </script>
-  <!-- <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/latest/css/bootstrap.min.css"> -->
-  <link rel="stylesheet" href="bootstrap.min.css">
-</head>
-
-<body>
-<div id="root"></div>
-</body>
-
+  <head>
+    <style type="text/css">
+      body {
+        padding: 10px;
+        width: 500px;
+        height: 500px;
+      }
+    </style>
+    <script async src="popup.js">
+    </script>
+    <!-- <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/latest/css/bootstrap.min.css"> -->
+    <link rel="stylesheet" href="bootstrap.min.css">
+  </head>
+  <body>
+    <div id="root"></div>
+  </body>
 </html>
