+++ client/src/options_custom/index.html
@@ -5,43 +5,43 @@
 -->
 <!DOCTYPE html>
 <html>
-<head>
-  <meta charset="utf-8">
-  <title id="title"></title>
+  <head>
+    <meta charset="utf-8">
+    <title id="title"></title>
 
-  <!-- Stylesheets -->
-  <link id="favicon" rel="icon" href="">
-  <link rel="stylesheet" href="lib/default.css" media="screen">
-  <link rel="stylesheet" href="css/main.css" media="screen">
-  <link rel="stylesheet" href="css/setting.css" media="screen">
-  <link rel="stylesheet" href="custom.css" media="screen">
+    <!-- Stylesheets -->
+    <link id="favicon" rel="icon" href="">
+    <link rel="stylesheet" href="lib/default.css" media="screen">
+    <link rel="stylesheet" href="css/main.css" media="screen">
+    <link rel="stylesheet" href="css/setting.css" media="screen">
+    <link rel="stylesheet" href="custom.css" media="screen">
 
-  <!-- JavaScripts -->
-  <script src="lib/mootools-core.js"></script>
-  <script src="lib/store.js"></script>
-  <script src="js/classes/tab.js"></script>
-  <script src="js/classes/setting.js"></script>
-  <script src="js/classes/search.js"></script>
-  <script src="js/classes/fancy-settings.js"></script>
-  <script src="i18n.js"></script>
-  <script src="js/i18n.js"></script>
-  <script src="manifest.js"></script>
-  <script src="settings.js"></script>
-</head>
-<body class="no-select">
-<div id="sidebar" class="fancy">
-  <img id="icon" src="" alt="">
-  <h1 id="settings-label"></h1>
-  <div id="tab-container">
-    <div id="search-container" class="tab">
-      <input id="search" type="search" placeholder="">
+    <!-- JavaScripts -->
+    <script src="lib/mootools-core.js"></script>
+    <script src="lib/store.js"></script>
+    <script src="js/classes/tab.js"></script>
+    <script src="js/classes/setting.js"></script>
+    <script src="js/classes/search.js"></script>
+    <script src="js/classes/fancy-settings.js"></script>
+    <script src="i18n.js"></script>
+    <script src="js/i18n.js"></script>
+    <script src="manifest.js"></script>
+    <script src="settings.js"></script>
+  </head>
+  <body class="no-select">
+    <div id="sidebar" class="fancy">
+      <img id="icon" src="" alt="">
+      <h1 id="settings-label"></h1>
+      <div id="tab-container">
+        <div id="search-container" class="tab">
+          <input id="search" type="search" placeholder="">
+        </div>
+      </div>
     </div>
-  </div>
-</div>
-<div id="content">
-  <div id="search-result-container" class="tab-content">
-    <h2 id="search-label"></h2>
-  </div>
-</div>
-</body>
+    <div id="content">
+      <div id="search-result-container" class="tab-content">
+        <h2 id="search-label"></h2>
+      </div>
+    </div>
+  </body>
 </html>
