+++ component/client/core/ClientChat.tsx
@@ -1,7 +1,8 @@
 import * as React from "react";
 import {Panel} from "react-bootstrap";
-import Messenger from "../chat/Messenger";
+
 import ChatHistory from "../chat/ChatHistory";
+import Messenger from "../chat/Messenger";
 
 class ClientChat extends React.Component<any, any> {
   private messenger: JSX.Element;
@@ -9,7 +10,14 @@ class ClientChat extends React.Component<any, any> {
   constructor(props: any) {
     super(props);
 
-    this.messenger = <Messenger />;
+    // By react conventions, history should be instantiated in the constructor.
+    this.state = {
+      history: [],
+      pushMessage: ""
+    };
+
+    this.handleSend = this.handleSend.bind(this);
+    this.messenger = <Messenger handleSend={this.handleSend}/>;
   }
 
   public render() {
@@ -24,11 +32,15 @@ class ClientChat extends React.Component<any, any> {
       <div>
         <Panel header="Chat" bsStyle="primary" footer={this.messenger}>
           <div style={bodyStyle} >
-            <ChatHistory username={this.props.username} />
+            <ChatHistory username={this.props.username} url={this.props.url} pushMessage={this.state.pushMessage} />
           </div>
         </Panel>
       </div>
     );
+  }
+
+  private handleSend(message: string) {
+    this.setState({pushMessage: message});
   }
 }
