--- component/main.js
@@ -1,5 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import ClientCore from "./client/core/ClientCore.jsx";
-
-ReactDOM.render(<ClientCore />, document.getElementById("root"));
