+++ component/main.tsx
@@ -1,9 +1,8 @@
 import {initializeApp} from "firebase";
 import * as React from "react";
 import * as ReactDOM from "react-dom";
-import {initializeApp}  from "firebase";
 
-import ClientCore from "./client/ClientCore";
+import ClientCore from "./client/core/ClientCore";
 
 // Initialize Firebase
 const config = {
@@ -13,16 +12,6 @@ const config = {
   messagingSenderId: "1089725560944",
   projectId: "url-connet",
   storageBucket: "url-connet.appspot.com"
-};
-initializeApp(config);
-
-const config = {
-  apiKey: "AIzaSyBakHeV8lMlysuBRtIWU9vz_hv6dF_zHxM",
-  authDomain: "url-connet.firebaseapp.com",
-  databaseURL: "https://url-connet.firebaseio.com",
-  projectId: "url-connet",
-  storageBucket: "url-connet.appspot.com",
-  messagingSenderId: "1089725560944"
 };
 initializeApp(config);
